requirements.txt
astroid==2.1.0
Django==2.1.5
django-adminlte2==0.4.1
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.2.2
pytz==2018.9
six==1.12.0
typed-ast==1.2.0
wrapt==1.11.1
